# schemas/openapi-v1alpha1.json
{
  "openapi": "3.1.0",
  "info": {"title":"DARKSTAR local API","version":"1.0.0-alpha.1"},
  "servers": [{"url":"http://127.0.0.1:{port}","variables":{"port":{"default":"0"}}}],
  "security": [{"bearerAuth":[]}],
  "paths": {
...
    "/api/v1/projects/{projectId}/backlog":{"get":{"operationId":"getProjectBacklog","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"maximum":1000}},{"name":"cursor","in":"query","schema":{"type":"string"}},{"name":"includePrevious","in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":"Versioned read-only backlog observations and source configuration","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BacklogView"}}}},"default":{"$ref":"#/components/responses/Error"}}}},
...
    "/api/v1/projects/{projectId}/tickets":{"get":{"operationId":"listNativeTickets","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"q","in":"query","schema":{"type":"string"}},{"name":"state","in":"query","schema":{"type":"string"}},{"name":"cursor","in":"query","schema":{"type":"string"}},{"name":"pageSize","in":"query","schema":{"type":"integer","minimum":1,"maximum":100}}],"responses":{"200":{"description":"Native business tickets with preserved history","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NativeTicketPage"}}}},"default":{"$ref":"#/components/responses/Error"}}}},
    "/api/v1/projects/{projectId}/tickets/{ticketId}":{"get":{"operationId":"getNativeTicket","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"ticketId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Native business tickets with preserved history","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NativeTicketDetail"}}}},"default":{"$ref":"#/components/responses/Error"}}}},
...
    "/api/v1/projects": {
      "get": {"operationId":"listProjects","responses":{"200":{"description":"Registered project projections","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Project"}}}}},"default":{"$ref":"#/components/responses/Error"}}},
      "post": {"operationId":"registerProject","parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectRegistration"}}}},"responses":{"201":{"description":"Project registered","headers":{"Location":{"schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"default":{"$ref":"#/components/responses/Error"}}}
    },
...
    "/api/v1/logs/{reference}": {
      "get": {"operationId":"readLog","parameters":[{"$ref":"#/components/parameters/LogReference"},{"$ref":"#/components/parameters/LogAfter"},{"$ref":"#/components/parameters/LogLimit"}],"responses":{"200":{"description":"Bounded log bytes","headers":{"X-Darkstar-Log-Offset":{"schema":{"type":"integer","minimum":0}},"X-Darkstar-Log-Next-Offset":{"schema":{"type":"integer","minimum":0}},"X-Darkstar-Log-Size":{"schema":{"type":"integer","minimum":0}},"X-Darkstar-Log-Complete":{"schema":{"type":"boolean"}}},"content":{"application/octet-stream":{"schema":{"type":"string","format":"binary"}}}},"416":{"description":"Log cursor is beyond the current size","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"default":{"$ref":"#/components/responses/Error"}}}
    },
...
    "/api/v1/agents/{attemptId}/logs": {
      "get": {"operationId":"readAgentLog","parameters":[{"$ref":"#/components/parameters/AttemptId"},{"$ref":"#/components/parameters/LogAfter"},{"$ref":"#/components/parameters/LogLimit"}],"responses":{"200":{"description":"Bounded redacted provider log bytes","headers":{"X-Darkstar-Log-Offset":{"schema":{"type":"integer","minimum":0}},"X-Darkstar-Log-Next-Offset":{"schema":{"type":"integer","minimum":0}},"X-Darkstar-Log-Size":{"schema":{"type":"integer","minimum":0}},"X-Darkstar-Log-Complete":{"schema":{"type":"boolean"}}},"content":{"application/octet-stream":{"schema":{"type":"string","format":"binary"}}}},"default":{"$ref":"#/components/responses/Error"}}}
    },
...
  },
  "components": {
    "securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"}},
    "parameters": {
...
      "LogReference":{"name":"reference","in":"path","required":true,"schema":{"type":"string","pattern":"^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$"}},
      "LogAfter":{"name":"after","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0}},
      "LogLimit":{"name":"limit","in":"query","required":false,"schema":{"type":"integer","minimum":1,"maximum":1048576,"default":65536}},
...
      "AttemptId":{"name":"attemptId","in":"path","required":true,"schema":{"type":"string","pattern":"^attempt_[0-9A-HJKMNP-TV-Z]{26}$"}},
...
      "IdempotencyKey":{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","minLength":8,"maxLength":128}},
...
    },
    "responses":{"Error":{"description":"Stable API error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}},
    "schemas": {
...
      "BacklogNamespace":{"type":"object","additionalProperties":false,"required":["provider","host","tenantId","scopeId"],"properties":{"provider":{"type":"string"},"host":{"type":"string"},"tenantId":{"type":"string"},"scopeId":{"type":"string"}}},
      "BacklogScope":{"type":"object","additionalProperties":false,"required":["namespace","containerId"],"properties":{"namespace":{"$ref":"#/components/schemas/BacklogNamespace"},"containerId":{"type":"string"}}},
      "BacklogTicketRef":{"type":"object","additionalProperties":false,"required":["namespace","id"],"properties":{"namespace":{"$ref":"#/components/schemas/BacklogNamespace"},"id":{"type":"string"}}},
      "BacklogSource":{"oneOf":[{"type":"object","additionalProperties":false,"required":["kind"],"properties":{"kind":{"const":"built_in"},"namespace":{"$ref":"#/components/schemas/BacklogNamespace"}}},{"type":"object","additionalProperties":false,"required":["kind","connectionId","connectionRevision","scope"],"properties":{"kind":{"const":"external"},"connectionId":{"type":"string"},"connectionRevision":{"type":"string"},"scope":{"$ref":"#/components/schemas/BacklogScope"}}}]},
      "BacklogBinding":{"type":"object","additionalProperties":false,"required":["projectId","revision","source","selectedAt"],"properties":{"projectId":{"type":"string"},"revision":{"type":"integer","minimum":0},"source":{"$ref":"#/components/schemas/BacklogSource"},"selectedAt":{"type":"string"}}},
...
      "BacklogPredicate":{"type":"object","additionalProperties":false,"required":["fieldId","operator","values"],"properties":{"fieldId":{"type":"string"},"operator":{"enum":["equals","in","after"]},"values":{"type":"array","items":{"type":"string"}}}},
      "BacklogQuery":{"type":"object","additionalProperties":false,"required":["text","predicates","pageSize"],"properties":{"text":{"type":"string"},"predicates":{"type":"array","items":{"$ref":"#/components/schemas/BacklogPredicate"}},"pageSize":{"type":"integer","minimum":1,"maximum":100}}},
...
      "BacklogFailure":{"type":"object","additionalProperties":false,"required":["code","message","retryable","details"],"properties":{"code":{"type":"string"},"message":{"type":"string"},"retryable":{"type":"boolean"},"details":{"type":"object","additionalProperties":{"type":"string"}}}},
      "BacklogRefresh":{"type":"object","additionalProperties":false,"required":["bindingRevision","phase","generation","query","startedAt","updatedAt","lastSuccessAt","nextAttemptAt","failures"],"properties":{"bindingRevision":{"type":"integer","minimum":0},"phase":{"enum":["refreshing","complete","failed"]},"generation":{"type":"integer","minimum":0},"query":{"$ref":"#/components/schemas/BacklogQuery"},"startedAt":{"type":"string"},"updatedAt":{"type":"string"},"lastSuccessAt":{"type":"string"},"nextAttemptAt":{"type":"string"},"failures":{"type":"integer","minimum":0},"error":{"$ref":"#/components/schemas/BacklogFailure"}}},
...
      "BacklogNamedID":{"type":"object","additionalProperties":false,"required":["id","name"],"properties":{"id":{"type":"string"},"name":{"type":"string"}}},
      "BacklogNamedObservation":{"oneOf":[{"type":"object","additionalProperties":false,"required":["state","value"],"properties":{"state":{"const":"known"},"value":{"$ref":"#/components/schemas/BacklogNamedID"}}},{"type":"object","additionalProperties":false,"required":["state","reason"],"properties":{"state":{"const":"unknown"},"reason":{"type":"string"}}},{"type":"object","additionalProperties":false,"required":["state","reason"],"properties":{"state":{"const":"unsupported"},"reason":{"type":"string"}}}]},
      "BacklogNamesObservation":{"oneOf":[{"type":"object","additionalProperties":false,"required":["state","value"],"properties":{"state":{"const":"known"},"value":{"type":"array","items":{"$ref":"#/components/schemas/BacklogNamedID"}}}},{"type":"object","additionalProperties":false,"required":["state","reason"],"properties":{"state":{"const":"unknown"},"reason":{"type":"string"}}},{"type":"object","additionalProperties":false,"required":["state","reason"],"properties":{"state":{"const":"unsupported"},"reason":{"type":"string"}}}]},
      "BacklogBooleanObservation":{"oneOf":[{"type":"object","additionalProperties":false,"required":["state","value"],"properties":{"state":{"const":"known"},"value":{"type":"boolean"}}},{"type":"object","additionalProperties":false,"required":["state","reason"],"properties":{"state":{"const":"unknown"},"reason":{"type":"string"}}},{"type":"object","additionalProperties":false,"required":["state","reason"],"properties":{"state":{"const":"unsupported"},"reason":{"type":"string"}}}]},
      "BacklogTicket":{"type":"object","additionalProperties":false,"required":["ref","ticketKey","observationId","bindingRevision","currentSource","currentQueryMatch","revision","key","url","title","description","status","reason","observedAt","checkedAt","evidenceRef","businessState","priority","assignees","labels","archived","freshness"],"properties":{"ref":{"$ref":"#/components/schemas/BacklogTicketRef"},"ticketKey":{"type":"string"},"observationId":{"type":"string"},"bindingRevision":{"type":"integer","minimum":0},"currentSource":{"type":"boolean"},"currentQueryMatch":{"type":"boolean"},"revision":{"type":"string"},"key":{"type":"string"},"url":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"status":{"enum":["cached","fresh","incomplete","missing","inaccessible","archived","out_of_scope"]},"reason":{"type":"string"},"observedAt":{"type":"string"},"checkedAt":{"type":"string"},"evidenceRef":{"type":"string"},"businessState":{"$ref":"#/components/schemas/BacklogNamedObservation"},"priority":{"$ref":"#/components/schemas/BacklogNamedObservation"},"assignees":{"$ref":"#/components/schemas/BacklogNamesObservation"},"labels":{"$ref":"#/components/schemas/BacklogNamesObservation"},"archived":{"$ref":"#/components/schemas/BacklogBooleanObservation"},"freshness":{"enum":["fresh","stale","never_observed"]}}},
      "BacklogView":{"type":"object","additionalProperties":false,"required":["schemaVersion","binding","query","refresh","tickets","nextCursor","includePrevious"],"properties":{"schemaVersion":{"const":1},"binding":{"$ref":"#/components/schemas/BacklogBinding"},"query":{"$ref":"#/components/schemas/BacklogQuery"},"refresh":{"anyOf":[{"$ref":"#/components/schemas/BacklogRefresh"},{"type":"null"}]},"tickets":{"type":"array","items":{"$ref":"#/components/schemas/BacklogTicket"}},"nextCursor":{"type":"string"},"includePrevious":{"type":"boolean"}}},
...
      "NativeTicketNamedID":{"type":"object","additionalProperties":false,"required":["id","name"],"properties":{"id":{"type":"string"},"name":{"type":"string"}}},
      "NativeTicket":{"type":"object","additionalProperties":false,"required":["id","projectId","revision","key","url","title","description","businessState","priority","assignees","labels","evidenceRef","observedAt"],"properties":{"id":{"type":"string"},"projectId":{"type":"string"},"revision":{"type":"string"},"key":{"type":"string"},"url":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"businessState":{"$ref":"#/components/schemas/NativeTicketNamedID"},"priority":{"type":"integer","minimum":0},"assignees":{"type":"array","items":{"$ref":"#/components/schemas/NativeTicketNamedID"}},"labels":{"type":"array","items":{"$ref":"#/components/schemas/NativeTicketNamedID"}},"evidenceRef":{"type":"string"},"observedAt":{"type":"string","format":"date-time"}}},
      "NativeTicketPage":{"type":"object","additionalProperties":false,"required":["schemaVersion","tickets","nextCursor"],"properties":{"schemaVersion":{"const":1},"tickets":{"type":"array","items":{"$ref":"#/components/schemas/NativeTicket"}},"nextCursor":{"type":"string"}}},
      "NativeTicketCapabilities":{"type":"object","additionalProperties":false,"required":["edit","transitions"],"properties":{"edit":{"type":"boolean"},"transitions":{"type":"boolean"}}},
      "NativeTicketField":{"type":"object","additionalProperties":false,"required":["id","name","kind","required"],"properties":{"id":{"type":"string"},"name":{"type":"string"},"kind":{"enum":["text","number","ids"]},"required":{"type":"boolean"}}},
      "NativeTicketTransition":{"type":"object","additionalProperties":false,"required":["id","name","toState"],"properties":{"id":{"type":"string"},"name":{"type":"string"},"toState":{"$ref":"#/components/schemas/NativeTicketNamedID"}}},
      "NativeTicketHistory":{"type":"object","additionalProperties":false,"required":["revision","kind","recordedAt","evidenceRef","title","description","businessState","priority"],"properties":{"revision":{"type":"string"},"kind":{"type":"string"},"recordedAt":{"type":"string","format":"date-time"},"evidenceRef":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"businessState":{"type":"string"},"priority":{"type":"integer","minimum":0}}},
      "NativeTicketDetail":{"type":"object","additionalProperties":false,"required":["schemaVersion","ticket","capabilities","fields","transitions","history"],"properties":{"schemaVersion":{"const":1},"ticket":{"$ref":"#/components/schemas/NativeTicket"},"capabilities":{"$ref":"#/components/schemas/NativeTicketCapabilities"},"fields":{"type":"array","items":{"$ref":"#/components/schemas/NativeTicketField"}},"transitions":{"type":"array","items":{"$ref":"#/components/schemas/NativeTicketTransition"}},"history":{"type":"array","items":{"$ref":"#/components/schemas/NativeTicketHistory"}}}},
...
      "ApiError":{"$ref":"./api-error-v1alpha1.schema.json"},
...
      "ProjectRegistration":{"type":"object","additionalProperties":false,"required":["name","source"],"properties":{"name":{"type":"string","minLength":1},"source":{"type":"string","minLength":1,"description":"Existing absolute project directory visible to the local daemon."}}},
      "Project":{"type":"object","additionalProperties":false,"required":["id","name","sourceHash","status","resourceVersion","lastGlobalPosition","createdAt","updatedAt"],"properties":{"id":{"type":"string","pattern":"^project_[0-9A-HJKMNP-TV-Z]{26}$"},"name":{"type":"string","minLength":1},"sourceHash":{"type":"string","pattern":"^[0-9a-f]{64}$"},"status":{"enum":["active","archived"]},"resourceVersion":{"type":"integer","minimum":1},"lastGlobalPosition":{"type":"integer","minimum":1},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},
...
    }
  }
}